pyproject.toml
[project]
name = "facial-emotion-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tensorflow<2.11",
    "numpy<2.0",
    "scikit-learn==1.2.2",
    "pandas==1.5.3",
    "matplotlib==3.7.5",
    "seaborn==0.12.2",
    "pillow==9.5.0",
    "tqdm==4.66.1",
    "ipykernel",
    "opencv-python",
    "streamlit",
    "gradio",
    "protobuf<3.20",
]

[tool.uv]
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
]
